requirements.txt
Flask==2.2.0
Werkzeug==2.2.2
joblib==1.0.0
pandas==1.4.2
numpy
matplotlib==3.3.4
scikit-learn==1.0.2
tensorflow==2.4.1
Pillow==8.4.0
gunicorn